requirements.txt
git+https://github.com/Rapptz/discord.py
python-dotenv
letterboxdpy
bs4
lxml
mysql.connector